Revert "Add flow to make a bet"
