+++ domain/src/core/entity.ts
@@ -1,9 +1,5 @@
 export type Id = `${string}-${string}-${string}-${string}-${string}`;
 
-export class Entity {
-  protected id: Id;
-
-  constructor(id: Id) {
-    this.id = id;
-  }
+export interface Entity {
+  id: Id;
 }
