+++ domain/src/features/match/index.ts
@@ -2,7 +2,3 @@ export { MatchEntity } from "./match.entity";
 export * from "./domain-events";
 export { GetMatchByIdParam } from "./usecases/get-match-by-id/get-match-by-id.param";
 export { GetMatchByIdUsecase } from "./usecases/get-match-by-id/get-match-by-id.usecase";
-export { MatchContract } from "./match.contract";
-export { CreateMatchParam } from "./usecases/create-match/create-match.param";
-export { CreateMatchUsecase } from "./usecases/create-match/create-match.usecase";
-
