--- domain/src/features/match/usecases/create-match/create-match.param.ts
@@ -1,5 +1,0 @@
-export interface CreateMatchParam {
-  matchId: string;
-  localTeam: string;
-  awayTeam: string;
-}
