--- domain/src/features/match/usecases/create-match/create-match.usecase.ts
@@ -1,4 +1,0 @@
-import { Usecase } from "../../../../core/usecase";
-import { CreateMatchParam } from "./create-match.param";
-
-export abstract class CreateMatchUsecase extends Usecase<CreateMatchParam> {}
