+++ domain/src/features/match/usecases/get-match-by-id/get-match-by-id.param.ts
@@ -1,3 +1,3 @@
-export interface GetMatchByIdParam {
-  matchId: string;
+export interface GetMatchByIdParam{
+    
 }
