+++ domain/src/features/team/team.entity.ts
@@ -1,14 +1,5 @@
-import { Entity, Id } from "../../core/entity";
+import { Entity } from "../../core/entity";
 
-export class TeamEntity extends Entity {
+export interface TeamEntity extends Entity {
   name: string;
-  
-  private constructor(id: Id, name: string) {
-    super(id);
-    this.name = name;
-  }
-  
-  build(params: { name: string }): TeamEntity {
-    return new TeamEntity(this.id, params.name);
-  }
 }
