+++ domain/src/features/user/index.ts
@@ -5,5 +5,3 @@ export { UserContract } from "./user.contract";
 
 export { GetUserByIdParam } from "./usecases/get-user-by-id/get-user-by-id.param";
 export { GetUserByIdUsecase } from "./usecases/get-user-by-id/get-user-by-id.usecase";
-export { CreateUserParam } from "./usecases/create-user/create-user.param";
-export { CreateUserUsecase } from "./usecases/create-user/create-user.usecase";
