--- domain/src/features/user/usecases/create-user/create-user.param.ts
@@ -1,3 +1,0 @@
-export interface CreateUserParam {
-  name: string;
-}
