--- domain/src/features/user/usecases/create-user/create-user.usecase.ts
@@ -1,4 +1,0 @@
-import { Usecase } from "../../../../core/usecase";
-import { CreateUserParam } from "./create-user.param";
-
-export abstract class CreateUserUsecase extends Usecase<CreateUserParam> {}
