+++ domain/src/features/user/user.entity.ts
@@ -1,14 +1,5 @@
-import { Entity, Id } from "../../core/entity";
+import { Entity } from "../../core/entity";
 
-export class UserEntity extends Entity {
+export interface UserEntity extends Entity {
   name: string;
-  
-  private constructor(id: Id, name: string) {
-    super(id);
-    this.name = name;
-  }
-  
-  build(params: { name: string }): UserEntity {
-    return new UserEntity(this.id, params.name);
-  }
 }
