+++ server/src/features/match/usecases/get-match-by-id.usecase-impl.ts
@@ -2,25 +2,15 @@ import {
   GetMatchByIdParam,
   GetMatchByIdUsecase,
   MatchDoesNotExistDomainEvent,
-  GottenMatchDomainEvent,
-  MatchContract
 } from "@skorify/domain/match";
 import { DomainEvent } from "@skorify/domain/core";
 
-export class GetMatchByIdUsecaseImpl extends GetMatchByIdUsecase {
-  constructor(private matchContract: MatchContract) {
+export class GetmatchByIdUsecaseImpl extends GetMatchByIdUsecase {
+  constructor() {
     super();
   }
 
   async call(param: GetMatchByIdParam): Promise<DomainEvent> {
-    const { matchId } = param;
-  
-      const matchInDB = await this.matchContract.getById(matchId);
-  
-      if (!matchInDB) {
-        return MatchDoesNotExistDomainEvent();
-      }
-  
-      return GottenMatchDomainEvent(matchInDB);
+    return MatchDoesNotExistDomainEvent();
   }
 }
