--- server/src/features/user/infrastructure/controller.ts
@@ -1,7 +1,0 @@
-import { IracaController } from "@scifamek-open-source/iraca/web-api";
-
-export class UserController extends IracaController {
-  configureEndpoints(): void {
-    this.configureEndpointsByPattern(/[\w]+Usecase$/, {});
-  }
-}
