+++ server/src/features/user/usecases/get-user-by-id.usecase-impl.ts
@@ -21,6 +21,9 @@ export class GetUserByIdUsecaseImpl extends GetUserByIdUsecase {
       return NotGottenUserDomainEvent();
     }
 
-    return GottenUserDomainEvent(userInDB);
+    return GottenUserDomainEvent({
+      id: "7e6f6d92-e8bd-4aab-a027-66f4ec31ba8e",
+      name: "Bryan",
+    });
   }
 }
